backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
django-filter==24.3
drf-yasg==1.21.8
channels==4.2.0
channels-redis==4.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.3.0
daphne==4.1.2
asgiref==3.8.1
PyJWT==2.10.1
cryptography==44.0.0
Pillow==11.0.0
